Update fixtures for new PSET prefix
